Improve padding on phone and added confirm dialog when clearing search history

+++ src/app/search-view/search-view.component.ts
@@ -96,8 +96,10 @@ export class SearchViewComponent {
   }
 
   clearHistory() {
-    this.certHistory = [];
-    this.localStorageService.removeData(SEARCH_HISTORY_KEY);
+    if (confirm("Clear search history?")) {
+      this.certHistory = [];
+      this.localStorageService.removeData(SEARCH_HISTORY_KEY);
+    }
   }
 
   removeHistoryEntry(index: number) {
